fix(poll): address security review findings in SocketPoll module
- Fix VALIDATE_MAXEVENTS signedness bug: check maxevents <= 0 before
  casting to size_t to properly detect negative values
- Add bounds check for kqueue ident cast: validate kev->ident fits in
  int before casting to prevent theoretical truncation on 64-bit systems
- Add hash chain length limit (SOCKET_MAX_HASH_CHAIN_LENGTH=16) to
  prevent algorithmic complexity attacks via hash collisions (DoS)

+++ include/poll/SocketPoll_backend.h
@@ -272,7 +272,8 @@ typedef struct PollBackend_T *PollBackend_T;
 #define VALIDATE_MAXEVENTS(maxevents, event_type)                             \
   do                                                                          \
     {                                                                         \
-      if ((size_t)(maxevents) <= 0)                                           \
+      /* Check before cast to detect negative values correctly */             \
+      if ((maxevents) <= 0)                                                   \
         {                                                                     \
           errno = EINVAL;                                                     \
           return NULL;                                                        \
